refactor(project): Fixes #72, pare back to initial skeletons
Pare back to original intent and get ready for a first public release

BREAKING CHANGE: Removes support for some skeletons. While some will make a return before the first
release, Parceljs and React will not.

fix #72
