+++ CONFIG-BUNDLES.md
@@ -95,37 +95,3 @@ npm i -D --save-exact babel-cli babel-preset-env eslint
 - [babel-cli](https://www.npmjs.com/package/babel-cli)
 - [babel-preset-env](https://github.com/babel/babel/tree/master/packages/babel-preset-env)
 - [eslint](https://www.npmjs.com/package/eslint)
-
-## simple-parcel
-
-TODO: Also remember to add instructions on editor config [https://prettier.io/docs/en/editors.html] - also add support for adding scripts to package.json
-
-This can either be used in conjunction with `common-web` or on it’s own if you project is purely JavaScript based.
-
-### config
-
-```
-"calavera": {
-    "commonjs": true
-}
-```
-
-### Skeletons
-
-This will add the following files to your project:
-
-- .eslintrc.js
-- .babelrc
-- .browserslistrc
-
-### Dependency installation
-
-When Calavera completes writing the above files, it will output the `npm` command you need to run in order to install the required dependencies. It will be of the form:
-
-```
-npm i -D --save-exact babel-cli babel-preset-env eslint
-```
-
-- [babel-cli](https://www.npmjs.com/package/babel-cli)
-- [babel-preset-env](https://github.com/babel/babel/tree/master/packages/babel-preset-env)
-- [eslint](https://www.npmjs.com/package/eslint)
