+++ closet/skeletons/.stylelintrc
@@ -1,76 +1,3 @@
 {
-    "rules": {
-        "at-rule-name-case": "lower",
-        "at-rule-name-space-after": "always-single-line",
-        "at-rule-semicolon-newline-after": "always",
-        "block-closing-brace-empty-line-before": "never",
-        "block-closing-brace-newline-after": "always",
-        "block-closing-brace-newline-before": "always",
-        "block-no-empty": true,
-        "block-opening-brace-newline-after": "always",
-        "block-opening-brace-space-before": "always",
-        "color-hex-case": "lower",
-        "color-named": "never",
-        "color-no-invalid-hex": true,
-        "comment-no-empty": true,
-        "declaration-bang-space-after": "never",
-        "declaration-bang-space-before": "always",
-        "declaration-block-no-shorthand-property-overrides": true,
-        "declaration-block-semicolon-newline-after": "always-multi-line",
-        "declaration-block-semicolon-newline-before": "never-multi-line",
-        "declaration-block-single-line-max-declarations": 1,
-        "declaration-block-trailing-semicolon": "always",
-        "declaration-colon-space-after": "always-single-line",
-        "declaration-colon-space-before": "never",
-        "declaration-empty-line-before": ["never", { ignore: ["after-comment", "after-declaration"] }],
-        "declaration-no-important": true,
-        "font-family-name-quotes": "always-where-recommended",
-        "font-family-no-duplicate-names": true,
-        "function-comma-space-after": "always-single-line",
-        "function-name-case": "lower",
-        "function-url-no-scheme-relative": true,
-        "function-url-quotes": "always",
-        "indentation": [4, { ignore: ["value"] }],
-        "keyframe-declaration-no-important": true,
-        "length-zero-no-unit": true,
-        "max-empty-lines": 2,
-        "max-nesting-depth": 5,
-        "media-feature-colon-space-after": "always",
-        "media-feature-colon-space-before": "never",
-        "media-feature-name-case": "lower",
-        "media-feature-name-no-unknown": [true, { ignoreMediaFeatureNames: ["min--moz-device-pixel-ratio"]}],
-        "media-feature-parentheses-space-inside": "never",
-        "media-feature-range-operator-space-after": "always",
-        "media-feature-range-operator-space-before": "always",
-        "media-query-list-comma-newline-after": "always-multi-line",
-        "media-query-list-comma-newline-before": "never-multi-line",
-        "media-query-list-comma-space-after": "always-single-line",
-        "media-query-list-comma-space-before": "never",
-        "no-eol-whitespace": true,
-        "no-extra-semicolons": true,
-        "no-invalid-double-slash-comments": true,
-        "no-missing-end-of-source-newline": true,
-        "no-unknown-animations": true,
-        "number-leading-zero": "never",
-        "number-no-trailing-zeros": true,
-        "property-case": "lower",
-        "property-no-unknown": true,
-        "rule-empty-line-before": ["always", { ignore: ["after-comment", "inside-block"] }],
-        "selector-combinator-space-after": "always",
-        "selector-combinator-space-before": "always",
-        "selector-list-comma-newline-after": "always",
-        "selector-list-comma-newline-before": "never-multi-line",
-        "selector-max-empty-lines": 0,
-        "selector-pseudo-class-case": "lower",
-        "selector-pseudo-class-no-unknown": true,
-        "selector-pseudo-element-case": "lower",
-        "selector-pseudo-element-no-unknown": true,
-        "selector-type-case": "lower",
-        "selector-type-no-unknown": true,
-        "string-no-newline": true,
-        "string-quotes": "single",
-        "unit-case": "lower",
-        "unit-no-unknown": true,
-        "value-keyword-case": "lower",
-    }
+    "extends": "stylelint-config-recommended"
 }
