--- closet/skeletons/react-parcel/.babelrc
@@ -1,13 +1,0 @@
-{
-    "presets": [
-        "@babel/react",
-        [
-            "@babel/env",
-            {
-                "targets": {
-                    "browsers": "last 2 versions"
-                }
-            }
-        ]
-    ]
-}
