--- closet/skeletons/react-parcel/.eslintrc.json
@@ -1,29 +1,0 @@
-{
-    "extends": [
-        "eslint:recommended",
-        "plugin:import/errors",
-        "plugin:react/recommended",
-        "plugin:jsx-a11y/recommended",
-        "prettier",
-        "prettier/react"
-    ],
-    "rules": {
-        "react/prop-types": 0,
-        "no-console": 1
-    },
-    "plugins": ["react", "import", "jsx-a11y"],
-    "parser": "babel-eslint",
-    "parserOptions": {
-        "ecmaVersion": 2018,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "jsx": true
-        }
-    },
-    "env": {
-        "browser": true,
-        "es6": true,
-        "jest": true,
-        "node": true
-    }
-}
