--- closet/skeletons/react-parcel/.prettierrc
@@ -1,1 +1,0 @@
-{}
