--- closet/skeletons/react-webpack/.babelrc
@@ -1,19 +1,0 @@
-{
-  "presets": [
-    "@babel/react",
-    [
-      "@babel/env",
-      {
-        "targets": {
-          "browsers": "last 2 versions"
-        },
-        "modules": false
-      }
-    ]
-  ],
-  "plugins": [
-    "react-hot-loader/babel",
-    "@babel/plugin-syntax-dynamic-import",
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
