--- closet/skeletons/simple-parcel/.eslintrc.json
@@ -1,12 +1,0 @@
-{
-  "extends": ["eslint:recommended", "prettier"],
-  "parserOptions": {
-    "ecmaVersion": 2018,
-    "sourceType": "module"
-  },
-  "env": {
-    "browser": true,
-    "es6": true,
-    "node": true
-  }
-}
