+++ lib/config-helpers/css.js
@@ -1,0 +1,10 @@
+const { writeFiles } = require("../fileUtils");
+
+cssTooling = {
+  addStylelint: () => {
+    writeFiles([".stylelintrc", ".stylelintignore"]);
+    return ["stylelint", "stylelint-config-recommended"];
+  }
+};
+
+module.exports = cssTooling;
