+++ lib/config-helpers/format.js
@@ -1,0 +1,10 @@
+const { writeFile } = require("../fileUtils");
+
+formatTooling = {
+  addPrettier: () => {
+    writeFile(".prettierrc");
+    return ["prettier"];
+  }
+};
+
+module.exports = formatTooling;
